// src/network/api/model/alarmResponse.ts
/**
 * Generated by orval v6.18.1 🍺
 * Do not edit manually.
 * Springdoc
 * Springdoc을 사용한 Swagger UI
 * OpenAPI spec version: 1.0.0
 */
import type { AlarmLocationResponse } from "./alarmLocationResponse";
import type { AlarmWeekResponse } from "./alarmWeekResponse";

export interface AlarmResponse {
  /** 알람ID */
  alarmId?: number;
  /** 활성화여부(0 : 비활성화, 1 : 활성화) */
  enabled?: number;
  /** 진동여부(0: 비활성화, 1 : 활성화) */
  vibration?: number;
  /** 알람시간(AM PM 시간) */
  alarmTime?: string;
  /** AM, PM */
  timeOfDay?: string;
  /** 알람지역목록 */
  alarmLocation?: AlarmLocationResponse[];
  /** 알람요일목록 */
  alarmWeek?: AlarmWeekResponse[];
}